requirements.txt
PyYAML
flask
gunicorn
python-dotenv